+++ apps/web/package.json
@@ -34,7 +34,7 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "compression": "^1.7.4",
-    "dayjs": "^1.11.12",
+    "date-fns": "^3.6.0",
     "express": "^4.18.2",
     "isbot": "^4.1.0",
     "lodash-es": "^4.17.21",
